fix(StatsBox): Moved to ResponsiveToggle. #312
